# Removed perf regression on format.Lookup