--- a/v2/binding/format/format.go
+++ b/v2/binding/format/format.go
@@ -46,9 +46,11 @@ func init() {
 
 // Lookup returns the format for contentType, or nil if not found.
 func Lookup(contentType string) Format {
-	if i := strings.IndexRune(contentType, ';'); i != -1 {
-		return formats[contentType[0:i]]
+	i := strings.IndexRune(contentType, ';')
+	if i == -1 {
+		i = len(contentType)
 	}
+	contentType = strings.TrimSpace(strings.ToLower(contentType[0:i]))
 	return formats[contentType]
 }